refactor: add service layer and switch prices to Decimal
- Introduce PriceService and route handlers call service instead of CRUD
- Use Decimal for price end-to-end (Deribit client -> CRUD -> ORM -> API schema)
- Update CRUD signatures and model typing to match Numeric(20, 8)
- Ignore celerybeat schedule shm/wal files

+++ app/db/crud.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import Mapping
 
 from sqlalchemy.orm import Session
@@ -7,12 +8,12 @@ from sqlalchemy.orm import Session
 from app.db.models import Price
 
 
-def save_price(session: Session, ticker: str, price: float, ts: int) -> None:
+def save_price(session: Session, ticker: str, price: Decimal, ts: int) -> None:
     row = Price(ticker=ticker, price=price, ts=ts)
     session.add(row)
 
 
-def save_prices(session: Session, prices: Mapping[str, float], ts: int) -> int:
+def save_prices(session: Session, prices: Mapping[str, Decimal], ts: int) -> int:
     """
     Сохраняет пачку цен за один timestamp.
     Возвращает количество добавленных строк.
