+++ app/db/models.py
@@ -1,4 +1,6 @@
-from sqlalchemy import BigInteger, Numeric, String, Index
+from decimal import Decimal
+
+from sqlalchemy import BigInteger, Index, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -10,7 +12,7 @@ class Price(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
 
     ticker: Mapped[str] = mapped_column(String(16), nullable=False)
-    price: Mapped[float] = mapped_column(Numeric(20, 8), nullable=False)
+    price: Mapped[Decimal] = mapped_column(Numeric(20, 8), nullable=False)
 
     ts: Mapped[int] = mapped_column(BigInteger, nullable=False)
