+++ app/schemas/price.py
@@ -1,9 +1,11 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class PriceOut(BaseModel):
     ticker: str
-    price: float
+    price: Decimal
     ts: int
 
     class Config:
